setup resource authorizaion

+++ app/controllers/projects_controller.rb
@@ -11,6 +11,7 @@ class ProjectsController < ApplicationController
 	
 	def create
 		create_project
+		assign_admin
 	end
 	
 	def update
@@ -24,34 +25,61 @@ class ProjectsController < ApplicationController
 	private
 	
 	def get_projects
-		projects = Project.all.order(created_at: :desc)
+		projects = @current_user.projects.order(created_at: :desc)
 		render json: projects, status: 200
 	end
 	
 	def get_project
 		project = Project.find(params[:id])
-		render json: project, status: 200
+		if project.users.include?(@current_user)
+			render json: project, status: 200
+		else
+			render json: { error: "unauthorized" }, status: 401
+		end
 	end
 	
 	def create_project
-		project = Project.create!(
+		@project = Project.create!(
 			title: 				params[:title],
 			description: 	params[:description]
 			)
-		render json: project, status: 200
+		render json: @project, status: 200
+	end
+	
+	def assign_admin
+		Subscription.create(
+			project_id: @project.id,
+			user_id: @current_user.id,
+			role: "owner"
+		)
+	end
+	
+	def check_owner
+		project = Project.find(params[:id])
+		sub = project.subscriptions.find_by(user_id: current_user.id)
+		return true if sub && sub[:role] == "owner"
+		return false
 	end
 	
 	def update_project
-		project = Project.find(params[:id])
-		project[:title] = params[:title]
-		project[:description] = params[:description]
-		project.save
-		render json: project, status: 200
+		if check_owner
+			project = Project.find(params[:id])
+			project[:title] = params[:title]
+			project[:description] = params[:description]
+			project.save
+			render json: project, status: 200
+		else
+			render json: { error: "unauthorized" }, status: 401
+		end
 	end
 	
 	def delete_project
-		project = Project.find(params[:id])
-		project.destroy!
-		render json: project, status: 200
+		if check_owner
+			project = Project.find(params[:id])
+			project.destroy!
+			render json: project, status: 200
+		else
+			render json: { error: "unauthorized" }, status: 401
+		end
 	end
 end
